requirements.txt
openai==0.27.6
line-bot-sdk==2.4.2